+++ src/components/produtos/formulario/formCadastro.tsx
@@ -1,4 +1,4 @@
-import { Button, ErrorAlert, SuccessAlert, TextField } from "@/components";
+import { Button, CadastroButtonModal, ErrorAlert, SuccessAlert, TextField } from "@/components";
 import { searchFornecedor } from "@/services/fornecedor/searchFornecedor";
 import axios from "axios";
 import { Formik, Form, Field } from "formik";
@@ -13,7 +13,7 @@ interface Fornecedor {
     telefone: string;
 }
 
-export const FormCadastroProduto: React.FC = () => {
+export const CadastroProdutoModal: React.FC = () => {
     const [isOpen, setIsOpen] = useState(false);
     const [SucessMessage, setSucessMessage] = useState("")
 
@@ -23,13 +23,11 @@ export const FormCadastroProduto: React.FC = () => {
 
     return (
         <div>
-            <button onClick={() => {setIsOpen(true)}} className="block text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800" type="button">
-                Novo produto
-            </button>
+            <CadastroButtonModal name="Novo produto" onClick={() => {setIsOpen(true)}} urlIcon="/icons/product-icon.svg" />
 
             {isOpen && (
                 <div className="fixed inset-0 z-50 flex items-center px-5 justify-center bg-black/20">
-                    <div className="bg-gray-50 rounded-2xl shadow-md px-4 py-6">
+                    <div className="w-full md:w-3xl bg-gray-50 rounded-2xl shadow-md px-6 py-8">
 
                         <Formik
                             initialValues={{
@@ -109,7 +107,7 @@ export const FormCadastroProduto: React.FC = () => {
                                         <ErrorAlert name="marca" component="div" />
                                     </div>
 
-                                    <div className="flex gap-3">
+                                    <div className="flex flex-col gap-3">
                                         <div>
                                             <TextField name="valor_unico" type="number" placeholder="Valor do produto" />
                                             <ErrorAlert name="valor_unico" component="div" />
@@ -121,7 +119,7 @@ export const FormCadastroProduto: React.FC = () => {
                                         </div>
                                     </div>
 
-                                    <div className="flex gap-3 items-center justify-center">
+                                    <div className="flex flex-col gap-3 items-center justify-center">
                                         <div className="w-full">
                                             <TextField type="text" name="fornecedor_nome" placeholder="Nome Fornecedor" />
                                             <ErrorAlert name="fornecedor_nome" component="div" />
